Settings: Port Lockscreen Targets and Shortcuts to SlimKat
First thanks to PA and CM which gave us partly code and some ideas.
Still some parts of the original implementation are/were buggy as well
we needed to ehance it.

Whatever the new Slim version is now ready and seems to be completly bug
free.

- Configurable lockscreen targets...on tablets 8 on phone either 5 or 8
- configurable lockscreen shortcuts only on phones...up to 5 shortcuts
- Centralazing some methods....using now then rounded corner on custom
  images systemwide...means on all other shorcut features we have
- multiDpi compatible
- new settings logic
- instance handling of all dialogs like we do on all other places
  way nicer this way
- tons of fixes

thanks kroz for redoing all xxhdpi icon in fb

Change-Id: I7749c592f75e51faa5db84d0a0ec314623c9b499

Conflicts:
	res/values/slim_arrays.xml
	src/com/android/settings/SecuritySettings.java
	src/com/android/settings/slim/util/ShortcutPickerHelper.java